backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
imagehash==4.3.1
anthropic==0.18.1
openai==1.12.0
python-dotenv==1.0.0
librosa==0.10.1
soundfile==0.12.1
pydantic==2.6.0
aiofiles==23.2.1
pedalboard>=0.9.0
scipy>=1.12.0
numpy>=1.26.0
moviepy>=2.0.0
elevenlabs>=1.0.0
